improved Responsiveness
Made the movie containers Grid displays to allow more movie cards display on each line. Fixed a bug where if you added a movie to the watch list and then searched for the same set of movies the add to watchlist button would allow you to add it again. Change the text inside the button to access the watchlist or search list screens depending on the screen you are on.

Still need to integrate the local storage feature and/or a database to remember the movies that have been added to the watchlist.